refactor: rename `lib` to `library`, use `into_owned` on `Cow` and add some comments
